// src/controllers/enums/WorkSessionEnum.ts
export enum WorkSessionEnum {
    WORK_SESSION_INVALIDE_ID = "ID de session invalide",
    WORK_SESSION_PAUSE = "Session introuvable ou déjà en pause",
    WORK_SESSION_RESUME = "Session introuvable ou non en pause",
    WORK_SESSION_NOT_FIND_PAUSE_SESSION = "Aucune pause en cours trouvée pour cette session",
    WORK_SESSION_STOP = "Session introuvable ou déjà terminée",
    WORK_SESSION_ALREADY_PAUSED = "Il y a déjà une pause en cours pour cette session",
    WORK_SESSION_ALREADY_ACTIVE = "Une session de travail est déjà en cours pour cet utilisateur"
}